feat(`server`): extract parseTRPCMessage for 3rd party modules
