+++ packages/server/src/rpc/index.ts
@@ -1,2 +1,3 @@
 export * from './codes';
 export * from './envelopes';
+export * from './parseTRPCMessage';
